Move candidate actions related to a folder and update import.
